Update how shells are defined
Instead of each shell being it's own .nix file, they are now defined as
attributes of the shells/default.nix file, which pulls in the
information for python and haskell shells currently as sub-attributes.
So, to launch a python34 shell, one would use:

    $ nix-shell ./shells -A python.python34

This also allows them to be inspected using nix-repl:

    $ nix-repl ./shells
    nix-repl> python.python34
    «derivation /nix/store/...-python3-3.4.3.drv»

I also cleaned things up a bit and broke the python.datastores
derivation (it actually needs to use python2) unless you make
modifications to the clouddbclient source. Hopefully that will be fixed
upstream soon enough.